requirements.txt
baidu-aip
pillow
selenium
pandas
openpyxl